Refactor: Change Action declarations to events in Enemy and Stage Lifecycle presenters and views; update uint to int for index variables

+++ Assets/InternalAssets/Scripts/Enemy/EnemyMovement/EnemyMovementPresenter.cs
@@ -38,7 +38,7 @@ public class EnemyMovementPresenter
     {
         if (_model.I < _model.Points.Count)
         {
-            _model.Point = _model.Points[(int)_model.I++];
+            _model.Point = _model.Points[_model.I++];
             
         }
         else
